% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{abstract.ts}
\alias{abstract.ts}
\title{Abstract Time Series}
\usage{
\method{abstract}{ts}(x, what = c("start", "end", "frequency"), ...)
}
\arguments{
\item{x}{the object to abstract}

\item{what}{which portion of the \code{ts} attribute to abstract, by default
all three are abstracted, but you can select, any one, two, or all}

\item{...}{arguments for methods that require further arguments}
}
\value{
a \code{ts} object with the \code{ts} parameter modified
}
\description{
\code{\link{alike}} will treat time series parameter components with zero in
them as wildcards.  This function allows you to create these wild card time
series attributes since R does not allow direct creation/modification of
\code{ts} attributes with zero values.
}
\details{
Make sure you do not try to use the templates you create with this for
anything other than as \code{\link{alike}} templates since the result is
likely undefined given R expects non zero values for the \code{ts}
attribute and attempts to prevent such attributes.
}
\seealso{
\code{\link{abstract}}
}
