% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-bw.R
\name{all_bw}
\alias{all_bw}
\title{Verify Values in Vector are Between Two Others}
\usage{
all_bw(x, lo, hi)
}
\arguments{
\item{x}{vector numeric or character}

\item{lo}{scalar vector of same type as \code{x}}

\item{hi}{scalar vector of same type as \code{y}}
}
\description{
Verify Values in Vector are Between Two Others
}
