% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{bench_mark}
\alias{bench_mark}
\title{Lightweight benchmarking function}
\usage{
bench_mark(..., times = 1000L)
}
\arguments{
\item{...}{expressions to benchmark, are captured unevaluated}

\item{times}{how many times to loop, defaults to 1000}
}
\value{
NULL, invisibly, reports timings as a side effect as screen output
}
\description{
Evaluates provided expression in a loop and reports mean evaluation time.
This is inferior to \code{microbenchmark} and other benchmarking tools in many
ways except that it has zero dependencies or suggests which helps with
package build and test times.  Used in vignettes.
}
\details{
Runs \code{\link[=gc]{gc()}} before first evaluation.
}
