% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstringr.R
\name{len_as_chr}
\alias{len_as_chr}
\title{Number Converted To character}
\usage{
len_as_chr(x)
}
\arguments{
\item{x}{integer(1L)}
}
\value{
character(1L)
}
\description{
e.g., \code{100} becomes \code{"100"}
}
