% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cstringr.R
\name{len_chr_len}
\alias{len_chr_len}
\title{Number of Characters Taken By Numeric Representation}
\usage{
len_chr_len(x)
}
\arguments{
\item{x}{integer(1L)}
}
\value{
integer(1L)
}
\description{
e.g., for 100, would be 3, for 1324, would be 4.
}
