% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{MachDblEpsSqrt}
\alias{MachDblEpsSqrt}
\title{Pre-calculated Precision Level}
\format{An object of class \code{numeric} of length 1.}
\usage{
MachDblEpsSqrt
}
\description{
Used to limit overhead of calls the require use of
\code{.Machine$double.eps ^ 0.5}
}
\keyword{datasets}
