% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{mk_df}
\alias{mk_df}
\title{Data Frame Constructor}
\usage{
mk_df(...)
}
\arguments{
\item{...}{the objects to use as columns for the data frame template}
}
\value{
a list classed as `data.frame`
}
\description{
A lighter weight constructor for data frames to use in \code{\link{validate}}.
Arguments are not checked or modified at all; constructing something that is
a reasonable template for a real data frame is solely the responsibility of
the user.
}
\note{
characters are not coerced to factors, etc.
}
\examples{
mk_df(a=character(), b=factor(levels=letters[1:6]))
}
