% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validators.R
\name{mk_val_token}
\alias{mk_val_token}
\title{Make Validator Token}
\usage{
mk_val_token(exp, err.msg = "")
}
\arguments{
\item{exp}{an expression which will be captured but not evaluated}

\item{err.msg}{character(1L) a message that tells the user what the
expected value should be}
}
\value{
a quoted expressions with \code{err.msg} attribute set
}
\description{
Utility function to generate validator tokens.
}
\details{
Allows you to supply error messages for validator to use for each error
token.  Your token should not contain top level \code{&&} or \code{||}.  If
it does your error message will not be reported.  If your token must involve
top level \code{&&} or \code{||}, use \code{identity(x && y)} to ensure that
your error message is used by \code{validate}.

Error messages should tell you what the acceptable values are so they mesh
with the rest of the error reporting messages.  Typically, we will append
"Should" ahead of your message, such that if you supply "not contain NAs"
it will be displayed as "Should not contain NAs".
}
\examples{
LEN.2 <- mk_val_token(length(.) == 2, "be length two")
}

