% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{name_sub}
\alias{name_sub}
\title{Manipulate Dot Only Names}
\usage{
name_sub(symb, arg_name)
}
\arguments{
\item{symb}{an R symbol}

\item{arg_name}{another R symbol}

\item{mode}{1L or 2L}
}
\value{
symbol
}
\description{
Used to support internal dot substitution and escaping
}
\keyword{internal}

