% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_validator}
\alias{parse_validator}
\title{Parse Expressions For \code{`validate`} Use}
\usage{
parse_validator(lang, arg_name, rho = parent.frame())
}
\arguments{
\item{arg_name}{another R symbol}

\item{symb}{an R symbol}

\item{an}{environment to look for language expressions to substitute}
}
\value{
list
}
\description{
Takes expressions provided to \code{`\link{validate}`} and identifies which
ones are templates, which ones are normal expressions, and substitutes the
actual argument name for \code{`.`}.
}
\details{
Internal function exposed for unit testing purposes
}
\keyword{internal}

