% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{symb_sub}
\alias{symb_sub}
\title{Expand variable names to experessions if they contain language}
\usage{
symb_sub(symb, env = parent.frame())
}
\description{
Expand variable names to experessions if they contain language
}
\keyword{internal}
