% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validators.R
\docType{data}
\name{validation_tokens}
\alias{CHR}
\alias{CHR1}
\alias{CPX}
\alias{CPX1}
\alias{INT}
\alias{INT.1}
\alias{LGL}
\alias{LGL1}
\alias{NUM}
\alias{NUM1}
\alias{validation_tokens}
\title{Atomic Vector validator}
\format{\preformatted{ language integer(1L) && NO.NA && NO.INF
}}
\usage{
INT.1
}
\description{
Atomic Vector validator
}
\seealso{
\code{\link{validator_sub}}
}
\keyword{datasets}

