% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetr-package.R
\docType{package}
\name{vetr-package}
\alias{vetr-package}
\title{Trust, but Verify}
\description{
Declarative template-based framework for verifying that objects meet
structural requirements, and auto-composing error messages when they do not.
}
