% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\docType{data}
\name{validation_tokens}
\alias{validation_tokens}
\alias{INT.1}
\alias{NUM.1}
\alias{CHR.1}
\alias{LGL.1}
\alias{CPX.1}
\alias{INT}
\alias{NUM}
\alias{CHR}
\alias{CPX}
\alias{LGL}
\title{Atomic Vector validator}
\format{An object of class \code{call} of length 3.}
\usage{
INT.1
}
\description{
Atomic Vector validator
}
\seealso{
\code{\link{validator_sub}}
}
\keyword{datasets}
