% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetr-package.R
\docType{package}
\name{vetr}
\alias{vetr}
\alias{vetr-package}
\title{vetr - Trust, but Verify}
\description{
A template based vetting mechanism for your function inputs that simplifies
argument checking for both complex and simple values while providing
informative messages to your users.
}
